Mirroring the HG repo - we forgot to for a while. A lot of changes in this one.
We're working on the help system and making it really, really spiffy. Right now,
our help files (accessible by `amp help [entry]`) are no longer hard-coded, but
either generated by command objects, or loaded from .markdown, .erb, or .txt files.
Also, markdown generates pretty (bold, underlined, sometimes colored) output!

Many of the commands have had some work done on them. The reason is because we
are moving our commands to the repository-independent API which we hope to
use to wrap our git support, darcs support, and so on.

Our functional tests no longer use the shell to test Amp by running it -
by modifying our dispatching code, we can now run tests in a tiny fraction
of the time. In addition, this gives us code coverage statistics, and a lot
of tests have been written.

We hope to have man pages ready, but we need to talk to the rubygems people to see
if it's possible to get them to install with our gem. If not, adgar has been working
on a spiffy man-related gem called rtfm, which we'll require as a dependency
with amp, making our man pages available to our users.

We've fixed hundreds of bugs. I'm not lying. Amp is getting better every day.